+++ packages/flutter_tts/tizen/src/text_to_speech.h
@@ -15,9 +15,8 @@
 enum class TtsState { kCreated, kReady, kPlaying, kPaused };
 
 using StateChangedCallback =
-    std::function<void(tts_state_e previous, tts_state_e current)>;
+    std::function<void(TtsState previous, TtsState current)>;
 using UtteranceCompletedCallback = std::function<void(int32_t utt_id)>;
-using ErrorCallback = std::function<void(int32_t utt_id, tts_error_e reason)>;
 
 class TextToSpeech {
  public:
@@ -35,25 +34,9 @@ class TextToSpeech {
     utterance_completed_callback_ = callback;
   }
 
-  void SetErrorCallback(ErrorCallback callback) { error_callback_ = callback; }
-
-  void OnStateChanged(tts_state_e previous, tts_state_e current) {
-    SwitchVolumeOnStateChange(previous, current);
-    state_changed_callback_(previous, current);
-  }
-
-  void OnUtteranceCompleted(int32_t utt_id) {
-    utterance_completed_callback_(utt_id);
-    ClearUttId();
-  }
-
-  void OnError(int32_t utt_id, tts_error_e reason) {
-    error_callback_(utt_id, reason);
-  }
-
   std::string GetDefaultLanguage() { return default_language_; }
 
-  void SetDefaultLanguage(std::string language) {
+  void SetDefaultLanguage(const std::string &language) {
     default_language_ = language;
   }
 
@@ -61,7 +44,7 @@ class TextToSpeech {
 
   std::optional<TtsState> GetState();
 
-  bool AddText(std::string text);
+  bool AddText(const std::string &text);
 
   bool Speak();
 
@@ -81,7 +64,7 @@ class TextToSpeech {
   void Prepare();
   void RegisterCallbacks();
   void UnregisterCallbacks();
-  void HandleAwaitSpeakCompletion(tts_state_e previous, tts_state_e current);
+
   void ClearUttId() { utt_id_ = 0; }
 
   void SwitchVolumeOnStateChange(tts_state_e previous, tts_state_e current);
@@ -100,7 +83,6 @@ class TextToSpeech {
 
   StateChangedCallback state_changed_callback_;
   UtteranceCompletedCallback utterance_completed_callback_;
-  ErrorCallback error_callback_;
 };
 
 #endif  // FLUTTER_PLUGIN_TEXT_TO_SPEACH_H_